pyproject.toml
[tool.poetry]
name = "collector"
version = "0.1.0"
description = ""
authors = ["Eric Kostandyan <ekostandyan@rvision.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"
confluent-kafka = "^2.0.2"
loguru = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
